+++ lib/providers/postgres-admin.js
@@ -11,11 +11,11 @@ pool.on('error', (error, client) => {
 	process.exit(-1)
 })
 
-const postgres = {
+const postgresClient = {
 	pool,
 	query: (text, params) => {
 		return pool.query(text, params)
 	},
 }
 
-export default postgres
+export default postgresClient
