+++ pages/api/import/firebase-realtime/migration.js
@@ -1,9 +1,27 @@
-import postgres from 'lib/postgres'
+import postgres from 'lib/providers/postgres-admin'
 import {getUserExport} from 'lib/providers/firebase-admin'
-import {insertChannel, insertUserChannel, insertTrack, insertChannelTrack} from 'lib/queries'
+import {insertChannel, insertUserChannel, insertTrack, insertChannelTrack} from './queries'
+
+// Reset database for debugging. Doesn't touch auth users.
+// Because queries will fail if something exists with same ids.
+// which order tho? channeltrack, channels, tracks, userchannel? or channels user_channel tracks channel_track
+function resetDb() {
+	return postgres.query(`
+		DELETE FROM channels;
+		DELETE FROM user_channel;
+		DELETE FROM tracks;
+		DELETE FROM channel_track;
+	`)
+}
+// await postgres.query('DELETE FROM public.channel_track')
+// await postgres.query('DELETE FROM public.channels')
+// await postgres.query('DELETE FROM public.tracks')
+// await postgres.query('DELETE FROM public.user_channel')
 
 export async function migrate({userFirebase, userSupabase}) {
 	console.log('running')
+
+	// await resetDb()
 
 	// etch data to migrate from Firebase.
 	const {channel, tracks} = await getUserExport(userFirebase.uid)
@@ -28,7 +46,7 @@ export async function migrate({userFirebase, userSupabase}) {
 	return true
 }
 
-export async function runQueries({supabaseUserId, channel, tracks}) {
+async function runQueries({supabaseUserId, channel, tracks}) {
 	if (!supabaseUserId) throw Error('supabaseUserId is required')
 	if (!channel) throw Error('channel is required')
 
@@ -50,7 +68,15 @@ export async function runQueries({supabaseUserId, channel, tracks}) {
 
 	let newTracks
 	try {
+		// @todo: maybe here do not filter out tracks with no url, but default empty string?
+		// can user have track with no url? maybe used (by users) as a marker of sections?
+		// i guess yea. but then we'd have to support in player etc etc <- does it not skip next already?
+		// we can test import/export for a while anyways, not ready to launch <-- yep
+		// i think the tracks without url are corrupt db entries. we require url otherwise
 		const trackQueries = tracks.filter((t) => t.url).map((track) => postgres.query(insertTrack(track)))
+
+		// @todo: maybe find a way to make sure all tracks are inserted; if fail, insert none?
+		// maybe it is in the postgres.query settings, or these mutation/procedure or what not
 		const results = await Promise.all(trackQueries)
 		newTracks = results.map((result) => {
 			return {
@@ -59,6 +85,8 @@ export async function runQueries({supabaseUserId, channel, tracks}) {
 			}
 		})
 	} catch (err) {
+		// That is handled here I believe. Because of the promise all
+		// i am not sure, i think there is maybe something specific in sql world
 		throw Error(err)
 	}
 
@@ -71,27 +99,3 @@ export async function runQueries({supabaseUserId, channel, tracks}) {
 		throw Error(err)
 	}
 }
-
-// Converts the Firebase timestamps to what Postgres wants
-// new Date("1411213745028").toISOString()
-// ==> "2014-09-20T11:49:05.028Z"
-function toTimestamp(timestamp) {
-	return new Date(Number(timestamp)).toISOString()
-}
-
-// Returns a timestamp from a Postgres datetime
-// const getTime = () =>
-// 	postgres.query('SELECT now()').then((rows) => {
-// 		return new Date(rows[0].now).getTime()
-// 	})
-
-// Reset database for debugging. Queries will fail if something exists with same ids.
-// await postgres.query('DELETE FROM public.channel_track')
-// await postgres.query('DELETE FROM public.channels')
-// await postgres.query('DELETE FROM public.tracks')
-// await postgres.query('DELETE FROM public.user_channel')
-
-// const logs = {start: await getTime(), end: 0, duration: 0}
-// logs.end = getTime()
-// logs.duration = logs.end - logs.start
-// console.log(`Migration ended in ${logs.duration / 1000} seconds`)
